group can add/remove users, logger started

+++ acl.service/src/main/java/controllers/ResourceController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 
 import authentication.EncryptionEngine;
+import authentication.TokenAuthentication;
 import model.Resource;
 import services.ResourceService;
 
@@ -19,49 +20,39 @@ public class ResourceController {
     
 	@Autowired
 	ResourceService resourceService;
+	TokenAuthentication tokenauth = new TokenAuthentication();
 	
 	
 	//create new resources by sending post req on /resources with the json data of resource
 	@RequestMapping(method = RequestMethod.POST, consumes = "application/json")  
-    public void create(@RequestBody Resource resource){
-    		resourceService.createResource(resource.getrName(), resource.getrPermissions());
+    public void create(@RequestParam(value="token", required=true) String token, @RequestBody Resource resource){
+		if(tokenauth.checkToken(token) && tokenauth.checkAdmin(token)) {
+			resourceService.createResource(resource.getrName(), resource.getrPermissions());
+		}
     }
 	
 	//Return the resource info
     @RequestMapping(value = "/{rId}", method = RequestMethod.GET, produces = "application/json")  
-    public Resource resource(@PathVariable int rId){
-    	return resourceService.getById(rId);
+    public Resource resource(@RequestParam(value="token", required=true) String token, @PathVariable int rId){
+    	if(tokenauth.checkToken(token) && tokenauth.checkAdmin(token)) {
+    		return resourceService.getById(rId);
+		}
+    	return null;
     }
     
     //update resource
     @RequestMapping(value = "/{rId}", method = RequestMethod.PUT, consumes = "application/json")  
-    public void update(@PathVariable int rId,@RequestBody Resource resource){
-    	resourceService.updateResource(rId, resource.getrName(), resource.getrPermissions());
+    public void update(@RequestParam(value="token", required=true) String token, @PathVariable int rId,@RequestBody Resource resource){
+    	if(tokenauth.checkToken(token) && tokenauth.checkAdmin(token)) {
+    		resourceService.updateResource(rId, resource.getrName(), resource.getrPermissions());
+		}
     }
     
     //delete resource
     @RequestMapping(value = "/{rId}", method = RequestMethod.DELETE)  
-    public void delete(@PathVariable int rId){
-    	resourceService.deleteResource(rId);
+    public void delete(@RequestParam(value="token", required=true) String token, @PathVariable int rId){
+    	if(tokenauth.checkToken(token) && tokenauth.checkAdmin(token)) {
+    		resourceService.deleteResource(rId);
+		}
     }
-	
-//    @RequestMapping(value = "/addresource", method = RequestMethod.POST, produces = "application/json")  
-//    public ModelAndView addresource(@RequestParam(value="name", required=true) String name, @RequestParam(value="pass", required=true) String pass){  
-//        EncryptionEngine encryptionEngine = new EncryptionEngine();
-//    	System.out.println(name+pass);
-//    	System.out.println(encryptionEngine.encryptWithMD5("hello"));
-//        return new ModelAndView("redirect:/");
-//    }
-//    
-//    @RequestMapping(value = "/removeresource", method = RequestMethod.POST, produces = "application/json")  
-//    public ModelAndView removeuser(@RequestParam(value="resourceid", required=true) int rId){  
-//        resourceService.deleteResource(rId);
-//        return new ModelAndView("redirect:/");
-//    }
-//    
-//    @RequestMapping(value = "/updateresource", method = RequestMethod.POST, produces = "application/json")  
-//    public ModelAndView updateuser(@RequestParam(value="resourceid", required=true) String rid){  
-//        
-//        return new ModelAndView("redirect:/");
-//    }
 }
